build: 1.15.4

+++ src/Controller/ContentManagement/FileController.php
@@ -10,6 +10,7 @@ use EMS\CoreBundle\Service\FileService;
 use Exception;
 use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\File\Exception\AccessDeniedException;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -44,8 +45,11 @@ class FileController extends AbstractController
     /**
      * @Route("/admin/file/{id}/soft-delete" , name="ems_file_soft_delete", methods={"POST","HEAD"})
      */
-    public function softDeleteFileAction(string $id, FileService $fileService): Response
+    public function softDeleteFileAction(Request $request, string $id, FileService $fileService): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException($request->getPathInfo());
+        }
         $fileService->removeSingleFileEntity([$id]);
 
         return $this->redirectToRoute('ems_core_uploaded_file_index');
@@ -54,8 +58,11 @@ class FileController extends AbstractController
     /**
      * @Route("/admin/file/{id}/hard-delete" , name="ems_file_hard_delete", methods={"POST","HEAD"})
      */
-    public function hardDeleteFileAction(string $id, FileService $fileService): Response
+    public function hardDeleteFileAction(Request $request, string $id, FileService $fileService): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException($request->getPathInfo());
+        }
         $fileService->hardRemoveFiles([$id]);
 
         return $this->redirectToRoute('ems_core_uploaded_file_index');
